requirements.txt
Flask==0.12.3
pyserial
RPi.Gpio